refactor : Mutatable String Extension 구현 및 적용
